+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * Constraint
 
-  * Add methods for removing groups from NonColliding constraints. [#247](https://github.com/personalrobotics/aikido/pull/247)
+  * Added methods for removing groups from NonColliding constraints. [#247](https://github.com/personalrobotics/aikido/pull/247)
 
 * Planner